refactor(codegen): extract events into plugin folder
