+++ scripts/codegen/src/core/registry.ts
@@ -4,6 +4,7 @@ import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { defaultChildrenPlugin } from "../plugins/defaultChildren/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
 import { elementByPropPlugin } from "../plugins/elementByProp/index.ts";
+import { eventsPlugin } from "../plugins/events/index.ts";
 import { examplesPlugin } from "../plugins/examples/index.ts";
 import { formControlPlugin } from "../plugins/formControl/index.ts";
 import { htmlAttrsPlugin } from "../plugins/htmlAttrs/index.ts";
@@ -42,4 +43,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   polymorphicPlugin,
   elementByPropPlugin,
   rootElementStaticPlugin,
+  eventsPlugin,
 ];
